requirements.txt
Django==3.1.5
python-decouple==3.4
django-ckeditor==6.0.0
django-admin-interface==0.14.2
djangorestframework==3.12.2
django-cors-headers==3.6.0
requests==2.25.1
Pillow==8.1.0
